Some -Wsign-conversion fixes
* Change duk_bool_to to duk_small_uint_t from duk_small_int_t.  This may
  cause some sign warnings in calling code.

* Reject attempt to unpack an array-like value whose length is 2G or over;
  previously was not checked explicitly, and the length was cast to duk_idx_t
  with a sign change and the unpack would then later fail.  Now it fails with
  a clean RangeError.

* Add wrap check for Node.js Buffer.concat().

* API DUK_TYPE_xxx, DUK_TYPE_MASK_xxx, flag constants etc are now unsigned.

+++ examples/alloc-hybrid/duk_alloc_hybrid.c
@@ -105,7 +105,7 @@ void *duk_alloc_hybrid_init(void) {
 #if defined(DUK_ALLOC_HYBRID_DEBUG)
 		printf("Pool %d: size %ld, count %ld\n", i, (long) pool_sizes[i].size, (long) pool_sizes[i].count);
 #endif
-		total_size += pool_sizes[i].size * pool_sizes[i].count;
+		total_size += pool_sizes[i].size * (size_t) pool_sizes[i].count;
 		if (pool_sizes[i].size > max_size) {
 			max_size = pool_sizes[i].size;
 		}
@@ -127,7 +127,7 @@ void *duk_alloc_hybrid_init(void) {
 		pool_header *hdr = st->headers + i;
 
 		hdr->alloc_start = p;
-		hdr->alloc_end = p + pool_sizes[i].size * pool_sizes[i].count;
+		hdr->alloc_end = p + pool_sizes[i].size * (size_t) pool_sizes[i].count;
 		hdr->free = (pool_free_entry *) (void *) p;
 		hdr->size = pool_sizes[i].size;
 		hdr->count = pool_sizes[i].count;
